requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==22.1.0
html-testRunner==1.2.1
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
six==1.16.0
tomli==2.0.1
